fix(settlement): 정산 요약 주 단위 집계 및 일별 매출 조회 오류 수정

+++ mmg-admin-service/src/main/java/com/green/mmg/admin/settlement/dto/SettlementSummaryRes.java
@@ -6,8 +6,8 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 public class SettlementSummaryRes {
-    private Integer expectedAmount;    // 예상 정산 금액
-    private Integer completedAmount;   // 완료 정산 금액
+    private Long expectedAmount;        // 예상 정산 금액
+    private Long completedAmount;      // 완료 정산 금액
     private Long completedCount;       // 완료 정산 건수
     private Long pendingCount;         // 대기 중 건수 (가게)
     private Long pendingStoreCount;    // 가게 대기 건수
